Make the Chrome Secure Shell format runnable in JavaScript console
The current format for Chrome Secure Shell is rather incomplete. It's JSON
without surrounding `{}` and can't be used for automation as it is.

This proposed format can be pasted into JavaScript console (Chrome DevTools) of
Chrome Secure Shell's options page to directly update the options.

The `term_.prefs_.set()` API is based on:
https://chromium.googlesource.com/apps/libapps/+/master/nassh/doc/FAQ.md